test: fix hang of app/fiber_channel.test.lua test
During testing process:

  https://gitlab.com/tarantool/tarantool/-/jobs/815901562#L5257

found that test app/fiber_channel.test.lua hanged:

  --- app/fiber_channel.result		Wed Sep  9 08:15:40 2020
  +++ var/152_app/fiber_channel.result	Wed Oct 28 04:51:34 2020
  @@ -211,370 +211,3 @@
   - false
   ...
   ch:get(box.info.pid) == box.info.pid
  ----
  -- true
  -...

This issue stops the testing till the complete timeout of the test-run
will stop all the testing process. To avoid of it 'wait_cond()' routine
was added to 'ch:get()' command to set it's local timeout to avoid of
waiting for test-run global timeout.