test: fix hang of vinyl/select_consistency.* test
During testing process:

  https://gitlab.com/tarantool/tarantool/-/jobs/800863457#L5303

found that test vinyl/select_consistency.test.lua hanged:

  --- vinyl/select_consistency.result		Thu Jun 11 11:59:48 2020
  +++ var/152_vinyl/select_consistency.result	Tue Oct 20 09:13:02 2020
  @@ -137,16 +137,3 @@
   for i = 1, ch:size() do
       ch:get()
   end;
  ----
  -...
  -test_run:cmd("setopt delimiter ''");
  ----
  -- true
  -...

This issue stops the testing till the complete timeout of the test-run
will stop all the testing process. To avoid of it 'wait_cond()' routine
was added to 'ch:get()' command to set it's local timeout to avoid of
waiting for test-run global timeout.