relay: use verbose names for fibers
Usually we use _f postfix for fiber's loop functions
and using same postfix for the fiber instance itself
looks inconsistent.

Same time if we grep for "struct fibers" we see a number
of places where fiber instances are postfixed with _fiber.

Thus lets make the same in relay fiber code:

 - use explicit reader_fiber name for a reader
 - use relay_fiber name for the joint relay fiber
   which depends on the reader, moreover this explicit
   name allows to note that the reader cancels the bound
   fiber if error happens.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>