raft: move box_update_ro_summary to update trigger
box_update_ro_summary() was called from the basic Raft library,
making it depend on box. But it was called every time when Raft
state was changed and broadcasted. It means the same effect can be
achieved by updating RO summary from Raft state update trigger.

The patch does it, and now Raft code does not depend on box.h.

Part of #5303